+++ crypto/test/src/test/DESedeTest.cs
@@ -47,10 +47,10 @@ namespace Org.BouncyCastle.Tests
         {
             private byte[] seed =
         {
-            (byte)0xaa, (byte)0xfd, (byte)0x12, (byte)0xf6, (byte)0x59,
-            (byte)0xca, (byte)0xe6, (byte)0x34, (byte)0x89, (byte)0xb4,
-            (byte)0x79, (byte)0xe5, (byte)0x07, (byte)0x6d, (byte)0xde,
-            (byte)0xc2, (byte)0xf0, (byte)0x6c, (byte)0xb5, (byte)0x8f
+            0xaa, 0xfd, 0x12, 0xf6, 0x59,
+            0xca, 0xe6, 0x34, 0x89, 0xb4,
+            0x79, 0xe5, 0x07, 0x6d, 0xde,
+            0xc2, 0xf0, 0x6c, 0xb5, 0x8f
         };
 
             public override void NextBytes(
@@ -219,7 +219,7 @@ namespace Org.BouncyCastle.Tests
 
                 for (int i = 0; i != input.Length / 2; i++)
                 {
-                    bytes[i] = (byte)dIn.ReadByte();
+                    bytes[i] = dIn.ReadByte();
                 }
 //				dIn.readFully(bytes, input.Length / 2, bytes.Length - input.Length / 2);
                 int remaining = bytes.Length - input.Length / 2;
